discord.py version 2 update

+++ models.py
@@ -42,13 +42,10 @@ class CodeDatabaseModel:
         table (str): Name of the database table to interact with
 
     """
+
     def __init__(self, table=None):
 
-        self.AVAILABLE_ATTRIBUTES = [
-            "name",
-            "code",
-            "rank"
-        ]
+        self.AVAILABLE_ATTRIBUTES = ["name", "code", "rank"]
 
         dotenv.load_dotenv()
         self.name = None
@@ -60,7 +57,7 @@ class CodeDatabaseModel:
             user=os.getenv("POSTGRES_USER"),
             password=os.getenv("POSTGRES_PASSWORD"),
             host=os.getenv("POSTGRES_HOST"),
-            port=os.getenv("POSTGRES_PORT")
+            port=os.getenv("POSTGRES_PORT"),
         )
         self._cursor = self._conn.cursor()
 
@@ -83,16 +80,13 @@ class CodeDatabaseModel:
             raise Exception("Name and code are required attributes")
         self._cursor.execute(
             f"INSERT INTO {table} (user_id, name, code, rank) VALUES (%s, %s, %s, %s);",
-            (user_id, name, code, rank)
+            (user_id, name, code, rank),
         )
         self._conn.commit()
 
     def _get_entry(self, table, user_id):
 
-        self._cursor.execute(
-            f'SELECT * from {table} WHERE user_id=%s;',
-            (user_id,)
-        )
+        self._cursor.execute(f"SELECT * from {table} WHERE user_id=%s;", (user_id,))
         entry = self._cursor.fetchone()
         if entry:
             return entry
@@ -119,10 +113,7 @@ class CodeDatabaseModel:
             raise Exception(f"Entry for user_id {user_id} not found. Entry must be created first")
 
         set_values = ", ".join([f"{item[0]} = '{item[1].upper()}'" for item in list(kwargs.items())])
-        self._cursor.execute(
-            f'UPDATE {table} SET {set_values} WHERE user_id=%s;',
-            (user_id,)
-        )
+        self._cursor.execute(f"UPDATE {table} SET {set_values} WHERE user_id=%s;", (user_id,))
         self._conn.commit()
 
     def _delete_entry(self, table, user_id):
@@ -131,10 +122,7 @@ class CodeDatabaseModel:
         if not entry[1]:
             raise Exception(f"Entry for user_id {user_id} not found. Entry must be created first")
 
-        self._cursor.execute(
-            f'DELETE FROM {table} WHERE user_id=%s;',
-            (user_id,)
-        )
+        self._cursor.execute(f"DELETE FROM {table} WHERE user_id=%s;", (user_id,))
         self._conn.commit()
 
 
@@ -145,6 +133,7 @@ class DDRCode(CodeDatabaseModel):
     Args:
         user_id (str): Discord User ID to reference
     """
+
     def __init__(self, user_id=None):
         self.user_id = user_id
         super().__init__("ddr_codes")
@@ -165,9 +154,7 @@ class DDRCode(CodeDatabaseModel):
             None
         """
         if not name or len(name) > 8:
-            raise Exception(
-                "A dancer name with at most 8 characters is required when creating a new entry"
-            )
+            raise Exception("A dancer name with at most 8 characters is required when creating a new entry")
         if not code:
             raise Exception("A DDR code (####-####) is required when creating a new entry")
 
@@ -246,6 +233,7 @@ class DDRCode(CodeDatabaseModel):
         self.code = None
         self.rank = None
 
+
 class IIDXCode(CodeDatabaseModel):
     """
     Object representing a IIDX player
@@ -253,6 +241,7 @@ class IIDXCode(CodeDatabaseModel):
     Args:
         user_id (str): Discord User ID to reference
     """
+
     def __init__(self, user_id=None):
         self.user_id = user_id
         super().__init__("iidx_codes")
@@ -273,9 +262,7 @@ class IIDXCode(CodeDatabaseModel):
             None
         """
         if not name or len(name) > 6:
-            raise Exception(
-                "A DJ name with at most 6 characters is required when creating a new entry"
-            )
+            raise Exception("A DJ name with at most 6 characters is required when creating a new entry")
         if not code:
             raise Exception("A IIDX ID is required when creating a new entry")
